fix/The modal I was working on is fine, but I am working on the implementation of another with less complexity
