+++ src/containers/App.jsx
@@ -1,13 +1,17 @@
 import React from "react";
 import { GlobalStyle } from "../components/styles/globals";
-import Routes from "../routes";
+/* import Routes from "../routes";
+ */
+ import ModalDottedMenu from '../components/Modals/ModalDottedMenú/index'
+ 
 
 function App() {
   return (
     <div className="App">
       <GlobalStyle /> 
-      <Routes />
-    </div>
+       <ModalDottedMenu /> 
+ {/*       <Routes />
+ */}    </div>
   );
 }
